Upgrade Metrika structure
