+++ metrika/commands/base.py
@@ -1,0 +1,4 @@
+class CommandBase:
+
+    def __init__(self, sdk):
+        self.sdk = sdk
