+++ metrika/events/base.py
@@ -1,0 +1,4 @@
+class EventBase:
+
+    def __init__(self, sdk):
+        self.sdk = sdk
