+++ metrika/main.py
@@ -1,9 +1,7 @@
-import random
-import string
-from time import time
-
 from sdk.codexbot_sdk import CodexBot
-from config import APPLICATION_TOKEN, APPLICATION_NAME, DB, URL, SERVER
+from config import APPLICATION_TOKEN, APPLICATION_NAME, DB, SERVER
+from commands.help import CommandHelp
+from commands.start import CommandStart
 
 
 class Metrika:
@@ -15,8 +13,8 @@ class Metrika:
         self.sdk.log("Metrika module initialized")
 
         self.sdk.register_commands([
-            ('metrika_help', 'help', self.help),
-            ('metrika_start', 'start', self.start)
+            ('metrika_help', 'help', CommandHelp(self.sdk).help),
+            ('metrika_start', 'start', CommandStart(self.sdk).start)
         ])
 
         self.sdk.set_routes([
@@ -25,32 +23,7 @@ class Metrika:
 
         self.sdk.start_server()
 
-    #
-    #
-    # HELP
-    # todo: move to the class
-    #
-
-    async def help(self, payload):
-        self.sdk.log("/help handler fired with payload {}".format(payload))
-
-        message = "Модуль для работы с сервисом Яндекс.Метрика.\n\n" \
-                  "Чтобы начать, выполните команду /metrika_start."
-
-        await self.sdk.send_to_chat(
-            payload["chat"],
-            message
-        )
-
-    async def start(self, payload):
-        self.sdk.log("/start handler fired with payload {}".format(payload))
-
-        message = "Стартуем."
-
-        await self.sdk.send_to_chat(
-            payload["chat"],
-            message
-        )
 
 if __name__ == "__main__":
     metrika = Metrika()
+
